# internal/server/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Bongsu Vulnerability Management API
  version: 0.1.0
  description: >
    Vulnerability management and SBOM platform API. Provides host inventory,
    vulnerability scanning, CVE database management, RBAC, and installer
    endpoints.
  license:
    name: MIT
servers:
  - url: /
    description: Current server
security: []
paths:
...
  /api/hosts:
    get:
      operationId: listHosts
      tags: [hosts]
      summary: List all hosts
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - name: agent_status
          in: query
          schema:
            type: string
            enum: [online, stale, offline, unknown]
          description: Filter by agent status
        - name: inventory_status
          in: query
          schema:
            type: string
            enum: [healthy, degraded, stale, empty, none]
          description: Filter by inventory status
        - name: agent_version_state
          in: query
          schema:
            type: string
            enum: [current, outdated, unknown]
          description: Filter by agent version drift state
      responses:
        "200":
          description: List of hosts with vulnerability counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostWithVulns"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/vuln-trends:
    get:
      operationId: getVulnTrends
      tags: [trends]
      summary: Get vulnerability trend data over time
      security:
        - AdminKey: []
        - ViewerKey: []
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Number of days to look back
        - name: host_id
          in: query
          schema:
            type: string
          description: Filter trends to a specific host
      responses:
        "200":
          description: Vulnerability trend rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VulnTrendRow"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/vuln-trends/summary:
    get:
      operationId: getVulnTrendSummary
      tags: [trends]
      summary: Get vulnerability trend summary with comparison
      security:
        - AdminKey: []
        - ViewerKey: []
        - bearerAuth: []
      responses:
        "200":
          description: Vulnerability trend summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnTrendSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Phase 4b: Intelligence ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key

    bearerAuth:
      type: http
      scheme: bearer
      description: Session token

    AdminKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key

    AgentKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Agent API key (or admin key)

    ViewerKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Viewer API key (subject:key format)

    InstallToken:
      type: apiKey
      in: header
      name: X-Install-Token
      description: Install token

  # ── Reusable Parameters ──────────────────────────────────────────────
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    Forbidden:
      description: Forbidden
      content:
        text/plain:
          schema:
            type: string
            example: forbidden
...
  schemas:
...
    Host:
      type: object
      properties:
        id:
          type: string
        hostname:
          type: string
        ip_address:
          type: string
        os_name:
          type: string
        os_version:
          type: string
        kernel:
          type: string
        arch:
          type: string
        cpu_model:
          type: string
        cpu_cores:
          type: integer
        memory_mb:
          type: integer
          format: int64
        agent_version:
          type: string
        agent_token_set:
          type: boolean
        owner:
          type: string
        team:
          type: string
        environment:
          type: string
        criticality:
          type: string
        tags:
          type: string
        last_seen:
          type: string
          format: date-time
        agent_status:
          type: string
          enum: [online, stale, offline, unknown]
        last_seen_age_seconds:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    HostWithVulns:
      allOf:
        - $ref: "#/components/schemas/Host"
        - type: object
          properties:
            vuln_counts:
              type: object
              additionalProperties:
                type: integer
            active_vuln_counts:
              type: object
              additionalProperties:
                type: integer
            latest_inventory:
              $ref: "#/components/schemas/HostInventorySummary"
    HostInventorySummary:
      type: object
      properties:
        scan_id:
          type: string
        package_count:
          type: integer
        vuln_count:
          type: integer
        container_count:
          type: integer
        scan_status:
          type: string
        scanned_at:
          type: string
          format: date-time
...
    VulnTrendRow:
      type: object
      properties:
        date:
          type: string
          format: date
        total_vulns:
          type: integer
        critical_count:
          type: integer
        high_count:
          type: integer
        medium_count:
          type: integer
        low_count:
          type: integer
        exploited_count:
          type: integer
        overdue_count:
          type: integer
        new_count:
          type: integer
        fixed_count:
          type: integer
        host_count:
          type: integer
    VulnTrendSummary:
      type: object
      properties:
        current:
          $ref: "#/components/schemas/VulnTrendRow"
        previous:
          $ref: "#/components/schemas/VulnTrendRow"
        trend_direction:
          type: string
          enum: [up, down, flat]
        total_hosts:
          type: integer
        snapshot_host_count:
          type: integer

    # ── Phase 4b: Intelligence ───────────────────────────────────────────
...